fix: 파트 매핑 파일 위치 이동
